# swagger.yaml
openapi: 3.0.3
info:
  title: Aquaculture Monitoring System API
  description: |
    A comprehensive API for monitoring aquaculture pools and sensor readings.
    This system tracks various water quality parameters, fish behavior, and environmental conditions
    to help maintain optimal conditions for aquaculture operations.
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:3000/api
    description: Development server
paths:
...
  /bacteria-density-readings:
    get:
      tags:
        - Bacteria Density Readings
      summary: Get all bacteria density readings
      description: Retrieve bacteria density readings with optional filtering
      parameters:
        - name: pool_id
          in: query
          description: Filter by pool ID
          schema:
            type: integer
        - name: limit
          in: query
          description: Number of records to return
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          description: Number of records to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of bacteria density readings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BacteriaDensityReading'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Bacteria Density Readings
      summary: Create new bacteria density reading
      description: Add a new bacteria density reading for a pool
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BacteriaDensityReadingCreate'
      responses:
        '201':
          description: Bacteria density reading created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BacteriaDensityReading'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /bacteria-density-readings/{id}:
    get:
      tags:
        - Bacteria Density Readings
      summary: Get bacteria density readings by pool ID
      description: Retrieve all bacteria density readings for a specific pool
      parameters:
        - name: id
          in: path
          required: true
          description: Pool ID
          schema:
            type: integer
      responses:
        '200':
          description: List of bacteria density readings for the pool
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BacteriaDensityReading'
        '404':
          description: No bacteria density readings found for this pool
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Current Sensor Values endpoints
...
  /daily-sensor-summary:
    get:
      tags:
        - Daily Sensor Summary
      summary: Get daily sensor summaries
      description: Retrieve daily aggregated sensor data
      parameters:
        - name: pool_id
          in: query
          description: Filter by pool ID
          schema:
            type: integer
        - name: sensor_type
          in: query
          description: Filter by sensor type
          schema:
            type: string
            enum: [ph, ammonia, nitrite, nitrate, dissolved_oxygen, orp, salinity, temperature, turbidity, water_level, toc, fish_activity, feeding_response, bacteria_density]
        - name: start_date
          in: query
          description: Start date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          description: End date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - name: limit
          in: query
          description: Number of records to return
          schema:
            type: integer
            default: 100
        - name: offset
          in: query
          description: Number of records to skip
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: List of daily sensor summaries
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DailySensorSummary'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      tags:
        - Daily Sensor Summary
      summary: Create daily sensor summary
      description: Add a new daily sensor summary
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DailySensorSummaryCreate'
      responses:
        '201':
          description: Daily sensor summary created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailySensorSummary'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Daily sensor summary already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /daily-sensor-summary/{id}:
    get:
      tags:
        - Daily Sensor Summary
      summary: Get daily sensor summary by ID
      description: Retrieve a specific daily sensor summary
      parameters:
        - name: id
          in: path
          required: true
          description: Summary ID
          schema:
            type: integer
      responses:
        '200':
          description: Daily sensor summary details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DailySensorSummary'
        '404':
          description: Daily sensor summary not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Water Purity Readings endpoints
...
  /norms:
    get:
      tags:
        - Norms
      summary: Get sensor limits/norms
      description: Retrieve current sensor limits and thresholds
      responses:
        '200':
          description: Current sensor norms
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Norms'
        '404':
          description: Norms not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      tags:
        - Norms
      summary: Update sensor limits/norms
      description: Update sensor limits and thresholds
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/NormsUpdate'
      responses:
        '200':
          description: Norms updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Norms'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    Pool:
      type: object
      properties:
        pool_id:
          type: integer
          description: Unique identifier for the pool
        number_of_fish:
          type: integer
          description: Number of fish in the pool
        age_of_fish:
          type: integer
          description: Age of fish in days
        capacity_liters:
          type: number
          description: Pool capacity in liters
        name:
          type: string
          description: Pool name
        location:
          type: string
          description: Pool location
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
      required:
        - pool_id
        - capacity_liters
...
    BacteriaDensityReading:
      type: object
      properties:
        reading_id:
          type: integer
          description: Unique identifier for the reading
        pool_id:
          type: integer
          description: Pool identifier
        bacteria_density_cfu_ml:
          type: number
          description: Bacteria density in CFU/ml
        total_bacteria_count:
          type: integer
          description: Total bacteria count
        pathogenic_bacteria:
          type: number
          description: Pathogenic bacteria percentage
        beneficial_bacteria:
          type: number
          description: Beneficial bacteria percentage
        bacteria_type:
          type: string
          description: Type or species of bacteria
        reading_timestamp:
          type: string
          format: date-time
          description: Reading timestamp
        notes:
          type: string
          description: Additional notes
        Pool:
          $ref: '#/components/schemas/Pool'
      required:
        - reading_id
        - pool_id
        - bacteria_density_cfu_ml
        - reading_timestamp
    BacteriaDensityReadingCreate:
      type: object
      properties:
        pool_id:
          type: integer
          description: Pool identifier
        bacteria_density_cfu_ml:
          type: number
          description: Bacteria density in CFU/ml (0-10,000,000)
        total_bacteria_count:
          type: integer
          description: Total bacteria count
        pathogenic_bacteria:
          type: number
          description: Pathogenic bacteria percentage (0-100)
        beneficial_bacteria:
          type: number
          description: Beneficial bacteria percentage (0-100)
        bacteria_type:
          type: string
          description: Type or species of bacteria
        notes:
          type: string
          description: Additional notes
      required:
        - pool_id
        - bacteria_density_cfu_ml
...
    DailySensorSummary:
      type: object
      properties:
        summary_id:
          type: integer
          description: Unique identifier for the summary
        pool_id:
          type: integer
          description: Pool identifier
        sensor_type:
          type: string
          description: Type of sensor
          enum: [ph, ammonia, nitrite, nitrate, dissolved_oxygen, orp, salinity, temperature, turbidity, water_level, toc, fish_activity, feeding_response, bacteria_density]
        summary_date:
          type: string
          format: date
          description: Date of the summary
        min_value:
          type: number
          description: Minimum value for the day
        max_value:
          type: number
          description: Maximum value for the day
        avg_value:
          type: number
          description: Average value for the day
        reading_count:
          type: integer
          description: Number of readings for the day
      required:
        - summary_id
        - pool_id
        - sensor_type
        - summary_date
    DailySensorSummaryCreate:
      type: object
      properties:
        pool_id:
          type: integer
          description: Pool identifier
        sensor_type:
          type: string
          description: Type of sensor
          enum: [ph, ammonia, nitrite, nitrate, dissolved_oxygen, orp, salinity, temperature, turbidity, water_level, toc, fish_activity, feeding_response, bacteria_density]
        summary_date:
          type: string
          format: date
          description: Date of the summary (YYYY-MM-DD)
        min_value:
          type: number
          description: Minimum value for the day
        max_value:
          type: number
          description: Maximum value for the day
        avg_value:
          type: number
          description: Average value for the day
        reading_count:
          type: integer
          description: Number of readings for the day
      required:
        - pool_id
        - sensor_type
        - summary_date
...
    Norms:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for norms
        PhMax:
          type: number
          description: Maximum pH value
        PhMin:
          type: number
          description: Minimum pH value
        AmmoniaMax:
          type: number
          description: Maximum ammonia value
        AmmoniaMin:
          type: number
          description: Minimum ammonia value
        NitriteMax:
          type: number
          description: Maximum nitrite value
        NitriteMin:
          type: number
          description: Minimum nitrite value
        NitrateMax:
          type: number
          description: Maximum nitrate value
        NitrateMin:
          type: number
          description: Minimum nitrate value
        DissolvedOxygenMax:
          type: number
          description: Maximum dissolved oxygen value
        DissolvedOxygenMin:
          type: number
          description: Minimum dissolved oxygen value
        OrpMax:
          type: number
          description: Maximum ORP value
        OrpMin:
          type: number
          description: Minimum ORP value
        SalinityMax:
          type: number
          description: Maximum salinity value
        SalinityMin:
          type: number
          description: Minimum salinity value
        TemperatureMax:
          type: number
          description: Maximum temperature value
        TemperatureMin:
          type: number
          description: Minimum temperature value
        TurbidityMax:
          type: number
          description: Maximum turbidity value
        TurbidityMin:
          type: number
          description: Minimum turbidity value
        WaterLevelMax:
          type: number
          description: Maximum water level value
        WaterLevelMin:
          type: number
          description: Minimum water level value
        TocMax:
          type: number
          description: Maximum TOC value
        TocMin:
          type: number
          description: Minimum TOC value
        FishActivityMax:
          type: number
          description: Maximum fish activity value
        FishActivityMin:
          type: number
          description: Minimum fish activity value
        FeedingResponseMax:
          type: number
          description: Maximum feeding response value
        FeedingResponseMin:
          type: number
          description: Minimum feeding response value
        BacteriaDensityMax:
          type: number
          description: Maximum bacteria density value
        BacteriaDensityMin:
          type: number
          description: Minimum bacteria density value
        WaterPurityMax:
          type: number
          description: Maximum water purity value
        WaterPurityMin:
          type: number
          description: Minimum water purity value
    NormsUpdate:
      type: object
      properties:
        PhMax:
          type: number
          description: Maximum pH value
        PhMin:
          type: number
          description: Minimum pH value
        AmmoniaMax:
          type: number
          description: Maximum ammonia value
        AmmoniaMin:
          type: number
          description: Minimum ammonia value
        NitriteMax:
          type: number
          description: Maximum nitrite value
        NitriteMin:
          type: number
          description: Minimum nitrite value
        NitrateMax:
          type: number
          description: Maximum nitrate value
        NitrateMin:
          type: number
          description: Minimum nitrate value
        DissolvedOxygenMax:
          type: number
          description: Maximum dissolved oxygen value
        DissolvedOxygenMin:
          type: number
          description: Minimum dissolved oxygen value
        OrpMax:
          type: number
          description: Maximum ORP value
        OrpMin:
          type: number
          description: Minimum ORP value
        SalinityMax:
          type: number
          description: Maximum salinity value
        SalinityMin:
          type: number
          description: Minimum salinity value
        TemperatureMax:
          type: number
          description: Maximum temperature value
        TemperatureMin:
          type: number
          description: Minimum temperature value
        TurbidityMax:
          type: number
          description: Maximum turbidity value
        TurbidityMin:
          type: number
          description: Minimum turbidity value
        WaterLevelMax:
          type: number
          description: Maximum water level value
        WaterLevelMin:
          type: number
          description: Minimum water level value
        TocMax:
          type: number
          description: Maximum TOC value
        TocMin:
          type: number
          description: Minimum TOC value
        FishActivityMax:
          type: number
          description: Maximum fish activity value
        FishActivityMin:
          type: number
          description: Minimum fish activity value
        FeedingResponseMax:
          type: number
          description: Maximum feeding response value
        FeedingResponseMin:
          type: number
          description: Minimum feeding response value
        BacteriaDensityMax:
          type: number
          description: Maximum bacteria density value
        BacteriaDensityMin:
          type: number
          description: Minimum bacteria density value
        WaterPurityMax:
          type: number
          description: Maximum water purity value
        WaterPurityMin:
          type: number
          description: Minimum water purity value
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...